🚀feat(Dockerfile): Update Tailscale version
Update the Tailscale version in the Dockerfile from 1.46.1 to 1.48.1.

🔧chore(fly.toml): Update app configuration file

Update the generated fly.toml app configuration file for the user "vepeen" with the new timestamp (2023-08-23T03:47:42+07:00).

🔒fix(start.sh): Update Tailscale hostname and tags

Update the Tailscale hostname and advertise tags in the start.sh script to include the user's region (yusoofs-fly{FLY_REGION} to yusoofs-{FLY_REGION}) and remove the flyexit tag.

Tailscale is now ready to be deployed.